requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0
pandas==2.2.0
matplotlib==3.8.2
seaborn==0.13.2
python-dotenv==1.0.1
